toevoegen vluchtcode en het aanmaken van een vliegtuig ook veranderen type gewicht in bagage

+++ src/Bagage.java
@@ -1,6 +1,6 @@
 public class Bagage {
 
-    float gewicht;
+    int gewicht;
 
     public Bagage( int gewicht) {
